# example_responses/openapi.json
{"openapi":"3.1.0","info":{"title":"Catholic Bulletin API","description":"Read-only access to parsed Catholic parish bulletins. Designed for mobile and web app consumption.","version":"0.1.0"},"paths":{"/v1/parishes":{"get":{"tags":["parishes"],"summary":"List Parishes","operationId":"list_parishes_v1_parishes_get","parameters":[{"name":"active_only","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Active Only"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParishListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parishes/{parish_id}":{"get":{"tags":["parishes"],"summary":"Get Parish","operationId":"get_parish_v1_parishes__parish_id__get","parameters":[{"name":"parish_id","in":"path","required":true,"schema":{"type":"string","title":"Parish Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParishResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parishes/{parish_id}/bulletins/current":{"get":{"tags":["bulletins"],"summary":"Get Current Bulletin","operationId":"get_current_bulletin_v1_parishes__parish_id__bulletins_current_get","parameters":[{"name":"parish_id","in":"path","required":true,"schema":{"type":"string","title":"Parish Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bulletin"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parishes/{parish_id}/bulletins/{week_starting}":{"get":{"tags":["bulletins"],"summary":"Get Bulletin By Date","operationId":"get_bulletin_by_date_v1_parishes__parish_id__bulletins__week_starting__get","parameters":[{"name":"parish_id","in":"path","required":true,"schema":{"type":"string","title":"Parish Id"}},{"name":"week_starting","in":"path","required":true,"schema":{"type":"string","format":"date","title":"Week Starting"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Bulletin"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parishes/{parish_id}/today":{"get":{"tags":["views"],"summary":"Get Today","operationId":"get_today_v1_parishes__parish_id__today_get","parameters":[{"name":"parish_id","in":"path","required":true,"schema":{"type":"string","title":"Parish Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodayResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/parishes/{parish_id}/schedule":{"get":{"tags":["views"],"summary":"Get Schedule","operationId":"get_schedule_v1_parishes__parish_id__schedule_get","parameters":[{"name":"parish_id","in":"path","required":true,"schema":{"type":"string","title":"Parish Id"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":7,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["meta"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"Announcement":{"properties":{"title":{"type":"string","title":"Title","description":"Short headline, ideally under 60 chars."},"body":{"type":"string","title":"Body","description":"Cleaned-up announcement text. Strip decorative emojis but keep meaning."},"category":{"$ref":"#/components/schemas/AnnouncementCategory"},"event_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Event Date","description":"If this is a one-time event, when it happens."},"event_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Event Time"},"location":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location","description":"Free-form location string (e.g., '32 Prince St', 'Parish Hall')."},"contact_email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Contact Email"},"contact_phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Contact Phone"},"register_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Register Url"},"priority":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Priority","description":"1=must-see (this week's events, urgent), 10=evergreen filler. App sorts by this.","default":5}},"type":"object","required":["title","body","category"],"title":"Announcement"},"AnnouncementCategory":{"type":"string","enum":["event","schedule_change","sacramental","ministry","stewardship","safety","operational","other"],"title":"AnnouncementCategory"},"Bulletin":{"properties":{"parish":{"$ref":"#/components/schemas/Parish"},"liturgical_day":{"$ref":"#/components/schemas/LiturgicalDay"},"week_starting":{"type":"string","format":"date","title":"Week Starting","description":"The Sunday this bulletin is published for (or covers)."},"recurring_schedule":{"items":{"$ref":"#/components/schemas/RecurringSlot"},"type":"array","title":"Recurring Schedule"},"schedule_exceptions":{"items":{"$ref":"#/components/schemas/ScheduleException"},"type":"array","title":"Schedule Exceptions"},"mass_intentions":{"items":{"$ref":"#/components/schemas/MassIntention"},"type":"array","title":"Mass Intentions"},"announcements":{"items":{"$ref":"#/components/schemas/Announcement"},"type":"array","title":"Announcements"},"collections":{"items":{"$ref":"#/components/schemas/Collection"},"type":"array","title":"Collections"},"raw_text_sha256":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Raw Text Sha256","description":"Hash of the source text — useful to detect duplicate parses and stale data."},"parser_version":{"type":"string","title":"Parser Version","default":"1.0.0"},"parsed_at":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parsed At","description":"ISO 8601 timestamp of when the parse happened."}},"type":"object","required":["parish","liturgical_day","week_starting","recurring_schedule"],"title":"Bulletin","description":"The full structured representation of one parish bulletin."},"Collection":{"properties":{"location_id":{"type":"string","title":"Location Id"},"amount_usd":{"type":"number","title":"Amount Usd"},"week_of":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Week Of"}},"type":"object","required":["location_id","amount_usd"],"title":"Collection","description":"A line from the 'last week's collection' report."},"DatedServiceResponse":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"start_time":{"type":"string","title":"Start Time"},"end_time":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Time"},"kind":{"type":"string","title":"Kind"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"location_id":{"type":"string","title":"Location Id"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"},"is_exception":{"type":"boolean","title":"Is Exception"},"intentions":{"items":{"type":"string"},"type":"array","title":"Intentions"}},"type":"object","required":["date","start_time","end_time","kind","language","location_id","notes","is_exception","intentions"],"title":"DatedServiceResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Language":{"type":"string","enum":["en","es","it","vi","la","pl","pt","fr","tl","ko","zh","other"],"title":"Language","description":"ISO-ish language codes for Mass language. Add more as encountered."},"LiturgicalDay":{"properties":{"name":{"type":"string","title":"Name","description":"e.g., 'Sixth Sunday of Easter'."},"date":{"type":"string","format":"date","title":"Date"},"readings":{"items":{"type":"string"},"type":"array","title":"Readings","description":"Citations only — e.g., ['Acts 8:5-8, 14-17', '1 Peter 3:15-18', 'John 14:15-21']. We don't store the full reading text — too much copyrighted material to redistribute."}},"type":"object","required":["name","date"],"title":"LiturgicalDay","description":"Liturgical identity of the Sunday/feast this bulletin covers."},"Location":{"properties":{"id":{"type":"string","title":"Id","description":"Stable identifier used to reference this location, e.g. 'main', 'mpb'."},"name":{"type":"string","title":"Name"},"address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Address"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"}},"type":"object","required":["id","name"],"title":"Location","description":"A worship site covered by this bulletin."},"MassIntention":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"time":{"type":"string","format":"time","title":"Time"},"intention_for":{"type":"string","title":"Intention For","description":"The person or cause the Mass is offered for, e.g. 'Mary & Anne De Bonis'."},"requested_by":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Requested By","description":"The person who requested the intention, if named."},"is_deceased":{"type":"boolean","title":"Is Deceased","description":"True for 'eternal rest of...', 'in memory of...' etc.","default":false},"location_id":{"type":"string","title":"Location Id","default":"main"}},"type":"object","required":["date","time","intention_for"],"title":"MassIntention","description":"A specific Mass said for a named intention on a specific date."},"Parish":{"properties":{"name":{"type":"string","title":"Name"},"locations":{"items":{"$ref":"#/components/schemas/Location"},"type":"array","title":"Locations"},"staff":{"items":{"$ref":"#/components/schemas/StaffMember"},"type":"array","title":"Staff"}},"type":"object","required":["name","locations"],"title":"Parish"},"ParishListResponse":{"properties":{"parishes":{"items":{"$ref":"#/components/schemas/ParishResponse"},"type":"array","title":"Parishes"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["parishes","count"],"title":"ParishListResponse"},"ParishResponse":{"properties":{"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"diocese":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Diocese"},"city":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"City"},"state":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"State"},"country":{"type":"string","title":"Country"},"timezone":{"type":"string","title":"Timezone"}},"type":"object","required":["id","name","diocese","city","state","country","timezone"],"title":"ParishResponse"},"RecurringSlot":{"properties":{"kind":{"$ref":"#/components/schemas/ServiceKind"},"weekday":{"$ref":"#/components/schemas/Weekday"},"start_time":{"type":"string","format":"time","title":"Start Time"},"end_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"End Time","description":"Only set when the bulletin specifies a window (typical for confessions, adoration)."},"language":{"anyOf":[{"$ref":"#/components/schemas/Language"},{"type":"null"}],"description":"Mass language. None when not specified or not applicable (e.g., confession)."},"location_id":{"type":"string","title":"Location Id","description":"References Location.id below. Use 'main' for single-location parishes."},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes","description":"Free-form context the parishioner should see — e.g., 'Vigil', 'Children's choir'."}},"type":"object","required":["kind","weekday","start_time","location_id"],"title":"RecurringSlot","description":"A regularly scheduled service — e.g., 'Sunday 10:30am English Mass'."},"ScheduleException":{"properties":{"kind":{"type":"string","enum":["added","cancelled","moved","modified"],"title":"Kind"},"date":{"type":"string","format":"date","title":"Date"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"For multi-day exceptions like 'temporary until Pentecost'."},"affects_service":{"$ref":"#/components/schemas/ServiceKind"},"affects_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"Affects Time","description":"The original time of the affected service, if applicable."},"new_time":{"anyOf":[{"type":"string","format":"time"},{"type":"null"}],"title":"New Time","description":"For 'moved' or 'added' exceptions."},"description":{"type":"string","title":"Description","description":"Human-readable explanation for the app to display."},"location_id":{"type":"string","title":"Location Id","default":"main"}},"type":"object","required":["kind","date","affects_service","description"],"title":"ScheduleException","description":"A one-off change to the regular schedule for a specific date.\n\nCaptures the things bulletins say in parenthetical asides:\n- 'No daily Mass Friday' -> kind=cancelled\n- 'temporary 5pm Mass until Pentecost' -> kind=added, with end_date\n- '12:10pm Mass moved to 7pm' -> kind=moved"},"ScheduleResponse":{"properties":{"parish_id":{"type":"string","title":"Parish Id"},"days":{"type":"integer","title":"Days"},"services":{"items":{"$ref":"#/components/schemas/DatedServiceResponse"},"type":"array","title":"Services"}},"type":"object","required":["parish_id","days","services"],"title":"ScheduleResponse"},"ServiceKind":{"type":"string","enum":["mass","vigil_mass","confession","adoration","rosary","benediction","novena","holy_hour","other"],"title":"ServiceKind"},"StaffMember":{"properties":{"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"}},"type":"object","required":["name","role"],"title":"StaffMember"},"TodayResponse":{"properties":{"parish_id":{"type":"string","title":"Parish Id"},"as_of":{"type":"string","format":"date-time","title":"As Of"},"today":{"type":"string","format":"date","title":"Today"},"next_service":{"anyOf":[{"$ref":"#/components/schemas/DatedServiceResponse"},{"type":"null"}]},"today_services_remaining":{"items":{"$ref":"#/components/schemas/DatedServiceResponse"},"type":"array","title":"Today Services Remaining"},"this_week_exceptions":{"items":{"$ref":"#/components/schemas/ScheduleException"},"type":"array","title":"This Week Exceptions"},"high_priority_announcements":{"items":{"$ref":"#/components/schemas/Announcement"},"type":"array","title":"High Priority Announcements"},"todays_intentions":{"items":{"$ref":"#/components/schemas/MassIntention"},"type":"array","title":"Todays Intentions"}},"type":"object","required":["parish_id","as_of","today","next_service","today_services_remaining","this_week_exceptions","high_priority_announcements","todays_intentions"],"title":"TodayResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"Weekday":{"type":"string","enum":["monday","tuesday","wednesday","thursday","friday","saturday","sunday"],"title":"Weekday"}}}}